feature not to install desktop apps

+++ entrypoint.sh
@@ -128,7 +128,7 @@ if [[ "${opt_nvidia_driver}" == "yes" ]]; then
   fi
 fi
 
-if [ ! -f tmp/step_2 ]; then
+if [[ ! "${desktop_apps}" == "no" ]] && [ ! -f tmp/step_2 ]; then
   echo 'execute step 3..'
   ./20_desktop-apps.sh 2>&1 | tee tmp/step_2.log && touch tmp/step_2
 fi
